catalog page finished, and details tho, need to do the carousel now, and to fix the equipment request + verification on the detail page (to not display the same car in the other cars category)

+++ rent-car/database/migrations/0001_01_01_000003_create_vehicles_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('brand', 100)->nullable(false);
             $table->string('model', 100)->nullable(false);
             $table->integer('year')->nullable(false);
-            $table->decimal('price_per_day', 10, 2)->nullable(false);
+            $table->integer('price_per_day')->nullable(false);
             $table->integer('doors')->nullable(false);
             $table->enum('fuel_type', ['essence', 'diesel', 'électrique', 'hybride'])->nullable(false);
             $table->boolean('air_conditioning')->default(false)->nullable(false);
